fix kimi default temperature

+++ src/BE/Services/Conversations/Implementations/Kimi/JsonKimiModelConfig.cs
@@ -8,5 +8,5 @@ public record JsonKimiModelConfig
     public required string Prompt { get; init; }
 
     [JsonPropertyName("temperature")]
-    public required double Temperature { get; init; }
+    public required float Temperature { get; init; }
 }
